# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.AttendanceData:
    properties:
      date_attendance:
        type: string
      id:
        type: integer
      status:
        type: string
    type: object
...
  dto.AttendanceRequest:
    properties:
      date:
        type: string
    required:
    - date
    type: object
  dto.AttendanceResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.AttendanceData'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total_pages:
        type: integer
      total_record:
        type: integer
    type: object
...
  dto.EmployeeResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
  dto.OvertimeData:
    properties:
      hourse:
        type: integer
      id:
        type: integer
      overtime_at:
        type: string
      status:
        type: string
    type: object
  dto.OvertimeRequest:
    properties:
      date:
        type: string
      hours:
        type: integer
    required:
    - date
    - hours
    type: object
  dto.OvertimeResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.OvertimeData'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total_pages:
        type: integer
      total_record:
        type: integer
    type: object
...
  dto.ReimbursementData:
    properties:
      amount:
        type: number
      description:
        type: string
      full_name:
        type: string
      id:
        type: integer
    type: object
  dto.ReimbursementRequest:
    properties:
      amount:
        type: integer
      date:
        type: string
      description:
        type: string
    required:
    - amount
    - date
    - description
    type: object
  dto.ReimbursementResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.ReimbursementData'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total_pages:
        type: integer
      total_record:
        type: integer
    type: object
  dto.RoleRequest:
    properties:
      name:
        type: string
      privilege:
        type: string
    required:
    - name
    - privilege
    type: object
  dto.RoleResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: API untuk manajemen payslip.
  title: Payslip API
  version: "1.0"
paths:
...
  /employee/attendance:
    post:
      consumes:
      - application/json
      description: Create a new attendance employee
      parameters:
      - description: attendance create payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AttendanceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.EmployeeResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create attendance employee
      tags:
      - employee
  /employee/attendance/all:
    get:
      consumes:
      - application/json
      description: Retrieves a list of attendacne with pagination.
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 20)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.AttendanceResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List of attendacne
      tags:
      - employee
  /employee/overtime:
    post:
      consumes:
      - application/json
      description: Create a new overtime employee
      parameters:
      - description: overtime create payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.OvertimeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.EmployeeResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create overtime employee
      tags:
      - employee
  /employee/overtime/all:
    get:
      consumes:
      - application/json
      description: Retrieves a list of overtime with pagination.
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 20)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.OvertimeResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List of overtime
      tags:
      - employee
  /employee/payslip/{id}:
    get:
      consumes:
      - application/json
      description: View detailed payslip for an employee for a specific payroll
      parameters:
      - description: Payroll ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.EmployeeResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get employee payslip
      tags:
      - employee
  /employee/reimbursement:
    post:
      consumes:
      - application/json
      description: Create a new reimbursement employee
      parameters:
      - description: reimbursement create payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ReimbursementRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.EmployeeResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create reimbursement employee
      tags:
      - employee
  /employee/reimbursement/all:
    get:
      consumes:
      - application/json
      description: Retrieves a list of reimbursement with pagination.
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 20)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ReimbursementResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List of reimbursement
      tags:
      - employee
  /roles/all:
    get:
      consumes:
      - application/json
      description: Get list of all roles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.RoleResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all roles
      tags:
      - roles
  /roles/create:
    post:
      consumes:
      - application/json
      description: Create a new roles
      parameters:
      - description: Roles create payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.RoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.RoleResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create roles
      tags:
      - roles
  /roles/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete roles
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.RoleResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete roles
      tags:
      - roles
  /roles/update/{id}:
    put:
      consumes:
      - application/json
      description: Update a new roles
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      - description: Roles update payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.RoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.RoleResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update roles
      tags:
      - roles
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"